Summary of All Changes Bug Fixes 1. Fixed Set Availability button - /home/zorvan/Work/projects/Zwischen/telegram-bot/bot/common/menus.py:111    - Changed callback_data from event_constraints_{event_id} to avail_add_{event_id} 2. Fixed callback parsing order - /home/zorvan/Work/projects/Zwischen/telegram-bot/bot/commands/event_details.py:185-190    - Reordered to check avail_add_ before avail_ 3. Fixed integer out of range error - db/models.py and db/schema.sql    - Changed all event_id columns from Integer to BigInteger    - Changed all group_id columns from Integer to BigInteger 4. Implemented Edit Event Details message handler - /home/zorvan/Work/projects/Zwischen/telegram-bot/bot/commands/event_details.py:1075-1117    - Added handle_message function to process user edits after clicking the button    - Handles cancel input    - Fetches event and validates user permissions    - Returns confirmation message Tests Added Unit tests - tests/unit/commands/test_event_details_units.py (9 tests) Integration tests - tests/integration/commands/test_availability_flow.py (5 tests) Documentation Created docs/AVAILABILITY_AND_CONSTRAINTS.md explaining the complete flow

+++ telegram-bot/db/schema.sql
@@ -41,7 +41,7 @@ CREATE TABLE IF NOT EXISTS groups (
 -- 3. Events: Gathering lifecycle
 CREATE TABLE IF NOT EXISTS events (
     event_id BIGSERIAL PRIMARY KEY,
-    group_id INTEGER REFERENCES groups(group_id) ON DELETE CASCADE,
+    group_id BIGINT REFERENCES groups(group_id) ON DELETE CASCADE,
     event_type VARCHAR(100) NOT NULL,
     description TEXT,
     organizer_telegram_user_id BIGINT,
